fix migrations for indexes

+++ backend/database/migrations/2019_09_08_013215_create_country_visitors_flow_index.php
@@ -8,6 +8,7 @@ class CreateCountryVisitorsFlowIndex extends Migration
 
     public function up()
     {
+        $this->down();
         $client = app('elasticsearch');
         $params = [
             'index' => self::INDEX_NAME,
@@ -25,10 +26,14 @@ class CreateCountryVisitorsFlowIndex extends Migration
         $client->indices()->create($params);
     }
 
+
     public function down()
     {
-        $client = app('elasticsearch');
         $params = ['index' => self::INDEX_NAME];
-        $client->indices()->delete($params);
+        $client = app('elasticsearch');
+
+        if ($client->indices()->exists($params)) {
+            $client->indices()->delete($params);
+        }
     }
 }
