+++ backend/database/migrations/2019_09_08_013620_create_device_visitors_flow_index.php
@@ -8,6 +8,7 @@ class CreateDeviceVisitorsFlowIndex extends Migration
 
     public function up()
     {
+        $this->down();
         $client = app('elasticsearch');
         $params = [
             'index' => self::INDEX_NAME,
@@ -28,6 +29,8 @@ class CreateDeviceVisitorsFlowIndex extends Migration
     {
         $client = app('elasticsearch');
         $params = ['index' => self::INDEX_NAME];
-        $client->indices()->delete($params);
+        if ($client->indices()->exists($params)) {
+            $client->indices()->delete($params);
+        }
     }
 }
